Dependency issue fixes
* Removed the requests library.

* Fixed the test discovery issue due to filename.

* Use std HTTP client for testing.
